Issue #6 - Handle Course Deleted Event
If a course is deleted, it first deletes all of the modules
inside that course. This triggers the syllabus_deleted event.

In that event, if a syllabus is deleted, a NULL entry is added
to demonstrate that the course has no syllabus resource. However,
if the entire course is then deleted, that means that we no longer
need to track this course, and we can delete all sv_entries.